Create basic M/M/c queue example
A simple M/M/c queue might be a nice addition to the examples. I'm not sure if this is the most straightforward way to deal with the arrival times. Maybe there's a way to pass an arrival distribution and a service distribution to customer (instead of an arrival time and a service distribution).