Allow priorities to be Floats
Extended the store and container objects so that priorities are not restricted to being integers.
This is a more general expression of priority which may make it easier at some point to use rank functions (https://github.com/BenLauwens/SimJulia.jl/issues/75).
This also avoids having to normalize (and sometimes scaling) priorities so that they are integer in queueing systems.